ToggleButtonBuilder ignores the value of isSelected() function
The problem was that the builder did not pass its isSelected attribute
over to DefaultToggleButton. In order to make it possible Selectable and
DefaultSelectable had to be modified as well. They now work the same way
as (for example) TextHolder and DefaultTextHolder: Selectable.create()
accepts a boolean argument that is passed over to DefaultSelectable's
constructor which sets its isSelected accordingly through the
selectedProperty.

The fix brought up another problem: the component was not rendered
properly and it looked like as if it wasn't "selected", however, the
internal isSelected attribute held the correct value. A TODO is added to
the DefaultToggleButton where it is supposed to be solved.

Fixes #187