Fixes #277 Hiding with transforming is leaving black background in LayerTransformerExample.
So the problem with this was that there is an optimization in the renderer which just skips rendering those tiles which would have been hidden below opaque tiles,
but this happened before applying TileTransformers. So if you added a TileTransformer which makes an opaque tile transparent to a tile anything below it was not
rendered at all!